bingo/backend/requirements.txt
bidict==0.22.1
blinker==1.7.0
click==8.1.7
dnspython==2.4.2
eventlet==0.33.3
Flask==3.0.0
flask-marshmallow==0.15.0
Flask-SocketIO==5.3.6
Flask-SQLAlchemy==3.1.1
greenlet==3.0.1
h11==0.14.0
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.3
marshmallow==3.20.1
marshmallow-sqlalchemy==0.29.0
mysql-connector==2.2.9
mysql-connector-python==8.2.0
mysqlclient==2.2.0
packaging==23.2
protobuf==4.21.12
python-dotenv==1.0.0
python-engineio==4.8.0
python-socketio==5.10.0
simple-websocket==1.0.0
six==1.16.0
SQLAlchemy==2.0.23
typing_extensions==4.8.0
Werkzeug==3.0.1
wsproto==1.2.0
